@@ -157,23 +151,18 @@
     "graphql": "16.5.0",
     "immer": "9.0.15",
     "io-ts": "2.2.18",
-    "libphonenumber-js": "1.10.6",
     "localforage": "1.10.0",
     "lodash": "4.17.21",
     "next": "12.3.6",
     "next-plausible": "3.12.4",
     "next-themes": "0.1.1",
     "next-transpile-modules": "9.1.0",
-    "node": "24.9.0",
-    "node-fetch": "npm:@blockworks-foundation/node-fetch@2.6.11",
-    "node-http": "0.0.5",
     "promptly": "3.2.0",
-    "protobufjs": ">=7.2.5",
     "psyfi-euros-test": "0.0.1-rc.33",
     "qr-code-styling": "1.6.0-rc.1",
     "ramda": "0.28.0",
     "rc-slider": "9.7.5",
-    "react": "18.2.0",
+    "react": "19.2.0",
     "react-async-hook": "4.0.0",
     "react-content-loader": "6.2.1",
     "react-dom": "18.0.0",
