@@ -2,9 +2,9 @@ import { PublicKey } from '@solana/web3.js'
 
 import { getRegistrarPDA, getVoterPDA } from 'VoteStakeRegistry/sdk/accounts'
 import {
-  ASSOCIATED_TOKEN_PROGRAM_ID,
+  ASSOCIATED_TOKEN_2022_PROGRAM_ID,
   Token,
-  TOKEN_PROGRAM_ID,
+  TOKEN_2022_PROGRAM_ID,
 } from '@solana/spl-token'
 import { VsrClient } from 'VoteStakeRegistry/sdk/client'
 import { TOKEN_2022_PROGRAM_ID } from '@solana/spl-token-new'
@@ -41,10 +41,10 @@ export const getClawbackInstruction = async ({
   const mintInfo = await client?.program.provider.connection.getAccountInfo(grantMintPk)
   const tokenProgram = mintInfo?.owner.equals(TOKEN_2022_PROGRAM_ID)
     ? TOKEN_2022_PROGRAM_ID 
-    : TOKEN_PROGRAM_ID
+    : TOKEN_2022_PROGRAM_ID
     
   const voterATAPk = await Token.getAssociatedTokenAddress(
-    ASSOCIATED_TOKEN_PROGRAM_ID,
+    ASSOCIATED_TOKEN_2022_PROGRAM_ID,
     tokenProgram,
     grantMintPk,
     voter,
