@@ -25,9 +25,9 @@ import {
 } from '@solana/web3.js'
 import { getATA } from '@utils/ataTools'
 import {
-  ASSOCIATED_TOKEN_PROGRAM_ID,
+  ASSOCIATED_TOKEN_2022_PROGRAM_ID,
   Token,
-  TOKEN_PROGRAM_ID,
+  TOKEN_2022_PROGRAM_ID,
 } from '@solana/spl-token'
 import { BN } from 'bn.js'
 
@@ -95,8 +95,8 @@ const ClaimUnderlyingPostExpiration = ({
       if (needToCreateAta) {
         prerequisiteInstructions.push(
           Token.createAssociatedTokenAccountInstruction(
-            ASSOCIATED_TOKEN_PROGRAM_ID,
-            TOKEN_PROGRAM_ID,
+            ASSOCIATED_TOKEN_2022_PROGRAM_ID,
+            TOKEN_2022_PROGRAM_ID,
             optionAccount.account.underlyingAssetMint,
             currentAddress,
             form.writerTokenAccount!.extensions.token!.account.owner,
@@ -116,7 +116,7 @@ const ClaimUnderlyingPostExpiration = ({
           form.writerTokenAccount.extensions.token!.account.address,
         underlyingAssetPool: optionAccount.account.underlyingAssetPool,
         underlyingAssetDest: underlyingDestination,
-        tokenProgram: TOKEN_PROGRAM_ID,
+        tokenProgram: TOKEN_2022_PROGRAM_ID,
         clock: SYSVAR_CLOCK_PUBKEY,
       },
     })
