@@ -1,7 +1,7 @@
 import {
-  ASSOCIATED_TOKEN_PROGRAM_ID,
+  ASSOCIATED_TOKEN_2022_PROGRAM_ID,
   Token,
-  TOKEN_PROGRAM_ID,
+  TOKEN_2022_PROGRAM_ID,
 } from '@solana/spl-token'
 import { TOKEN_2022_PROGRAM_ID } from "@solana/spl-token-new"
 import { PublicKey, TransactionInstruction } from '@solana/web3.js'
@@ -17,10 +17,10 @@ export const withCreateAssociatedTokenAccount = async (
     payerPk: PublicKey,
     useToken2022 = false,
 ) => {
-  const programId = useToken2022 ? TOKEN_2022_PROGRAM_ID : TOKEN_PROGRAM_ID
+  const programId = useToken2022 ? TOKEN_2022_PROGRAM_ID : TOKEN_2022_PROGRAM_ID
 
   const ataPk = await Token.getAssociatedTokenAddress(
-      ASSOCIATED_TOKEN_PROGRAM_ID,
+      ASSOCIATED_TOKEN_2022_PROGRAM_ID,
       programId,
       mintPk,
       ownerPk,
@@ -29,7 +29,7 @@ export const withCreateAssociatedTokenAccount = async (
 
   instructions.push(
       Token.createAssociatedTokenAccountInstruction(
-          ASSOCIATED_TOKEN_PROGRAM_ID,
+          ASSOCIATED_TOKEN_2022_PROGRAM_ID,
           programId,
           mintPk,
           ataPk,
