@@ -1,6 +1,6 @@
 // PATH: governance/tools/sdk/splToken/withMintTo.ts
 
-import {Token, TOKEN_PROGRAM_ID, u64} from '@solana/spl-token'
+import {Token, TOKEN_2022_PROGRAM_ID, u64} from '@solana/spl-token'
 import { TOKEN_2022_PROGRAM_ID } from '@solana/spl-token-new'
 import { PublicKey, TransactionInstruction } from '@solana/web3.js'
 
@@ -15,7 +15,7 @@ import { PublicKey, TransactionInstruction } from '@solana/web3.js'
  * @param destinationPk    Destination token account
  * @param mintAuthorityPk  Mint authority
  * @param amount           Amount to mint
- * @param useToken2022     If true, will use TOKEN_2022_PROGRAM_ID instead of TOKEN_PROGRAM_ID
+ * @param useToken2022     If true, will use TOKEN_2022_PROGRAM_ID instead of TOKEN_2022_PROGRAM_ID
  */
 // @ts-ignore: unused export
 /* eslint-disable @typescript-eslint/no-unused-vars */
@@ -27,7 +27,7 @@ export const withMintTo = async (
     amount: number | u64,
     useToken2022 = false, // <-- new param
 ) => {
-    const programId = useToken2022 ? TOKEN_2022_PROGRAM_ID : TOKEN_PROGRAM_ID
+    const programId = useToken2022 ? TOKEN_2022_PROGRAM_ID : TOKEN_2022_PROGRAM_ID
 
     instructions.push(
         Token.createMintToInstruction(
